semana03/dia4/requirements.txt
#framework principal
flask
#extensiones de flask

#otras librerias
mysql-connector-python